LayoutTests/fast/forms/input-text-default-border-radius-with-author-styled-border-expected.txt

This test checks that customizing the border of a text input removes any border radius set for the native appearance.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".



PASS getComputedStyle(input).borderRadius is "0px"
PASS successfullyParsed is true

TEST COMPLETE
